Add initial commit for docusaurus demo site
"Rough draft" demo using Docusaurus, so as to evaluate it as a
replacement for Sphinx. Once setup with GH pages, work on refining the
demo will continue.

This commit addresses FFTK 2415
This commit applies to FFTK 2414
This commit applies to FFTK 2312

QA Steps: few, but as of this commit results run in local dev server, as
well as build locally.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>