Add pdf files for offline use
Utilizing `https://github.com/jean-humann/docs-to-pdf`, we are able to
generate passable pdf versions of our documentation. There's many
options to fine tune the output, only a few were explored here.

There's a number of ways this could be integrated into the site, and
scripted for use during deployment. For the sake of the demo, things
were kept simple all around.

Images not appearing in the PDFs of v1.0.0 and v1.0.1 is the same as
with the versioned docs, which has a simple fix related to the asset
path. This issue has been kept around for the sake of being able to make
a note about it later during the write up.

This commit addresses FFTK-2739

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>